Refactor DeleteContainerAppStep to handle multiple container apps
