+++ src/commands/deleteManagedEnvironment/IDeleteManagedEnvironmentWizardContext.ts
@@ -4,14 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ExecuteActivityContext, IActionContext, ISubscriptionContext } from "@microsoft/vscode-azext-utils";
-import { ContainerAppTreeItem } from "../../tree/ContainerAppTreeItem";
 
 export interface IDeleteManagedEnvironmentWizardContext extends IActionContext, ExecuteActivityContext {
-    suppressPrompt?: boolean;
-
     subscription: ISubscriptionContext;
     resourceGroupName: string;
     managedEnvironmentName: string;
 
-    containerApps: ContainerAppTreeItem[];
+    containerAppNames: string[];
 }
